+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Controlled by `#thief:protected/<level>` entity-type tag.
   - By default, medium crime is commited when killing animals that are usually found in villages (Horses, Cows, Pigs, Chicken, Cats, etc).
   - Villagers are not included, as they are handled by vanilla already.
+- Added ability to show villager reputation tooltip without holding a gift item (config option). 
+- Fixed villager reputation tooltip rendering when gui is hidden. 
 
 ## 1.0.1 - 2025-05-24
 - Guard witnesses (Iron Golem, etc) will emit angry particles when crime is witnessed.